Bug fix: figured bass to read bass part ONLY
This fix assumes the bass part is the last part in the score.

Previously, every note in every part was being read sequentially. For example, in a two-part score with the quarter note 'E' in the upper part and quarter note 'C' in the lower part (both on beat 1), the module would have thought the bass line has 'E' on beat 1 and 'C' on beat 2, then erroneously realize that line.