Feature: passing/neighbour notes and rests
Takes care of passing/neighbour notes by asking whether it falls on a harmonically active beat, if not, then the previous chord will be held unless there is an explicit figured bass.

Since there is no longer a 1:1 correspondence between bass notes (indices) and chords in the right hand, the prior "hack solution" for rests no longer worked and required a complete overhaul of how to treats rests. They are now treated in place: the system looks ahead to the next chord to determine resolution. N.B.: only works with single rests in the middle of the piece at the moment – i.e., consecutive rests, rests at the beginning/end will not work.