Feature: figured bass extensions
Convention in figured bass is that a repeated figure is denoted with either a blank or an extender (as with lyrics). This update extracts the missing figure from the previous figure and appends it to the current figure. If the convention is not followed, then the extension may or may not be interpreted. For example, a "5,4" is followed by just a "3" with no blank is fine because the system understands "3" = "5,3" (as of the latest chord dictionary update).

This is compatible with the MusicXML 3.0 standard that denotes extensions with the <extend/> tag – by virtue of it being ignored, a blank is left within the 'figure' element, which is then understood as an extension.