Update: figured bass extensions
Previously, extensions were marked with empty lyric strings. The way the code ran places empty strings on every note unless stated, meaning that it was impossible to tell whether the empty strings meant an extension or a (5,3).

Now, extensions are marked using the attribute lyric.syllabic = 'end'. This required the creation of the method fbAppend() – a customized version of addElement() that doesn't reset the lyric attributes.

Figures tied over measures are handled (temporarily?) with the now-global variable prevFigure. This, however, doesn't work quite the same way with implied (5,3)'s, as the lack of marking means the chord will be realized again on beat 1 (which may be acceptable since it is on a strong beat). In general, a better parser for ties is needed.