Feature: diatonic seventh sequence (figured bass)
In a sequence of diatonic sevenths above descending fifths in the bass (also ascending fourths), standard stepwise resolutions result in complete and incomplete chords.

In minor keys, the melodic minor in the bass is assumed (i.e. natural 6 and 7) because the sequence is descending. Note: natural 7 results in a dominant seventh chord that – at this point – doesn't get treated properly because the original V7 method is still being called.