Feature (figured bass): 7-6 progressions and deceptive cadences
The following does not apply to regular dominant/diminished sevenths, which still follow the original code (will be merged later).

General seventh chords can now be automatically constructed either as complete or incomplete - a (7) is sufficient indication – and then resolved accordingly. When incomplete in four voices, the fifth is omitted and bass doubled.

This is necessary for the 7-6 progression which omits the fifth idiomatically (with no special indication other than what's on the score).

Deceptive cadences (V-vi or V-VI) are also added because they are easy to handle.