Bug fix: V6-I conditions (again)
If the first chord was the result of a V7/3-I6 resolution, there will be a doubled third. This means that this chord cannot become dominant harmony because of the doubled leading tone.

E.g. in the progression V7/3-I6-IV, the latter two chords cannot be interpreted as a "V6-I of IV."