Feature (figured bass): descending 6/3 sequence
Realizes the descending 6/3 sequence, where a stepwise-descending bass supports a sequence of 6/3 chords. To avoid parallel 5ths in four voices, the chords alternate doubling the bass (third) or the root.

Because the "resolution" is different depending on which note is doubled, this algorithm will first figure out which note is being doubled.