Bug fix (figured bass): rests compatibility
This was a "TO DO" for a very long time. Compatibility with rests for the :method:`getAllPossibilityProgressions' was not done along with the others because it was rarely called upon, and because its structure is different from the others - namely, it is non-random. Internally, the only use for it is when the requested number of solutions is greater than (or equal to) the actual number of solutions.