Feature (figured bass): right hand arpeggiator
A rudimentary arpeggiator is added as a proof-of-concept of a general purpose decorator. Arpeggiation patterns are prescribed in the algorithm for now but may be easily changed to a more complex algorithm.

How to use: after the realization object is created, change its :attr:`Realization.decorateKeyboardStyle' to {0 (default), 1, 2}. The arpeggiator will run when the user asks for a random possibility progression (i.e. produce a music21 stream.Score() object by calling either the :method:`generateRealizationFromPossibilityProgression' or the :method:`generateRandomRealization').